created a camera class
