+++ include/camera.h
@@ -1,0 +1,88 @@
+#ifndef CAMERA_H
+#define CAMERA_H
+
+#include "rtweekend.h"
+
+#include "Magick++.h"
+#include "hittable.h"
+
+#include <iostream>
+
+class camera {
+public:
+double aspect_ratio = 1.0;  // Ratio of image width over height
+int    image_width  = 100;  // Rendered image width in pixel count
+
+void render(const hittable& world) {
+    Magick::InitializeMagick("");
+    initialize();
+
+    Magick::Image image(Magick::Geometry(image_width, image_height), "white");
+
+    for (int j = 0; j < image_height; ++j) {
+        for (int i = 0; i < image_width; ++i) {
+            auto pixel_center = pixel00_loc + (i * pixel_delta_u) + (j * pixel_delta_v);
+            auto ray_direction = pixel_center - center;
+            ray r(center, ray_direction);
+
+            Magick::ColorRGB pixel_color = ray_color(r, world);
+            image.pixelColor(i, j, pixel_color);
+        }
+    }
+
+    image.write("./assets/images/sphere.png");
+
+}
+
+private:
+    int    image_height;   // Rendered image height
+    point3 center;         // Camera center
+    point3 pixel00_loc;    // Location of pixel 0, 0
+    vec3   pixel_delta_u;  // Offset to pixel to the right
+    vec3   pixel_delta_v;  // Offset to pixel below
+
+    void initialize() {
+        image_height = static_cast<int>(image_width / aspect_ratio);
+        image_height = (image_height < 1) ? 1 : image_height;
+
+        center = point3(0, 0, 0);
+
+        // Determine viewport dimensions.
+        auto focal_length = 1.0;
+        auto viewport_height = 2.0;
+        auto viewport_width = viewport_height * (static_cast<double>(image_width)/image_height);
+
+        // Calculate the vectors across the horizontal and down the vertical viewport edges.
+        auto viewport_u = vec3(viewport_width, 0, 0);
+        auto viewport_v = vec3(0, -viewport_height, 0);
+
+        // Calculate the horizontal and vertical delta vectors from pixel to pixel.
+        pixel_delta_u = viewport_u / image_width;
+        pixel_delta_v = viewport_v / image_height;
+
+        // Calculate the location of the upper left pixel.
+        auto viewport_upper_left =
+            center - vec3(0, 0, focal_length) - viewport_u/2 - viewport_v/2;
+        pixel00_loc = viewport_upper_left + 0.5 * (pixel_delta_u + pixel_delta_v);
+    }
+
+    Magick::ColorRGB background_color(const ray& r) {
+        vec3 unit_direction = unit_vector(r.direction());
+        auto a = 0.5*(unit_direction.y() + 1.0);
+        vec3 v = (1.0-a)*vec3(1.0, 1.0, 1.0) + a*vec3(0.5, 0.7, 1.0);
+
+        return Magick::ColorRGB(v.x(), v.y(), v.z());
+    }
+
+    Magick::ColorRGB ray_color(const ray& r, const hittable& world) {
+        hit_record rec;
+        if (world.hit(r, interval(0, infinity), rec)) {
+            vec3 color = 0.5 * (rec.normal + vec3(1,1,1));
+            return Magick::ColorRGB(color.x(), color.y(), color.z());
+        }
+
+        return background_color(r);
+    }
+};
+
+#endif
